Transaction: extracted interface Mappable #42
